+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0-pre.11] - 2021-11-5
+
+### Added
+
+- Add DictView to support dictionaries (lookup tables) in View
+- Add `ArrayView#indexOf` & `VectorView#indexOf` methods to support searching
+  items in container views.
+
 ## [4.0.0-pre.10] - 2021-10-20
 
 ### Added